feat: :rocket: se agrego el sistema de token para user

+++ src/controllers/user.js
@@ -1,20 +1,37 @@
 const md5 = require('md5');
 const scModel = require('../models/index');
 const hlpUser = require('../helpers/user');
+const service = require('../helpers/services');
+
 
 //global variables
 let user = {}
-let payload = {
-    message: null
+
+user.login = (req, res) =>{
+    try{
+        scModel.User.findOne({
+            "username": req.body.username
+        }).then(usr =>{
+            if(md5(req.body.password) == usr.password){
+                console.log("User login");
+                res.status(200)
+                res.send({message: "User login"})
+            }else{
+                console.log("password incorrect")
+                res.status(403)
+                res.send({message: "password incorrect"})
+            }
+        }).catch(err =>{
+            console.log("usuario no encontrado: ", err);
+            res.status(400);
+            res.send({message: "usuario no encontrado"})
+        })
+    }catch(err){
+        console.log("Error try:",err)
+    }
 };
 
-user.login = async (req, res, next ) =>{
-    res.send('login')
-    const pr = await scModel.User.find({"username": req.body.username});
-    console.log(pr);
-};
-
-user.signin = async (req, res , next) =>{
+user.signin = async (req, res) =>{
     var datosUser = req.body;
     try{
         var bandUser = await hlpUser.UserExist(req.body.username)
@@ -23,35 +40,56 @@ user.signin = async (req, res , next) =>{
             if(hlpUser.SamePass(req.body.password, req.body.rppassword)){
                 datosUser.password = md5(datosUser.password);
                 const newUser = new scModel.User(datosUser)
-                await newUser.save()
-                payload.message = 'User saved';
-                res.send(payload)
+                // eslint-disable-next-line no-unused-vars
+                await newUser.save(function(err){
+                    return res
+                    .status(200)
+                    .send({token: service.createToken(newUser)})
+                });
             }else{
-                payload.message = 'Contrasennas no similares';
-                res.send(payload)
+                res.status(403);
+                res.send({message: 'Contrasennas no similares'})
             }
         }else{
-            payload.message = 'username existente';
-            res.send(payload)
+            res.status(406)
+            res.send({message: 'username existente'})
         }
     }catch(err){
         console.log(err)
+        res.status(400)
         res.send(err)
     }
 };
 
-user.edit = async (req, res, next) => {
+user.edit = async (req, res) => {
     var datosUser = req.body
     datosUser.password = md5(datosUser.password);
     await scModel.User.findByIdAndUpdate(req.params.id, req.body);
-    payload.message = 'usuario actualizado';
-    res.send(payload);
+    res.status(200);
+    res.send({message: 'usuario actualizado'});
 }
 
-user.delete = async (req, res, next) =>{
-    await scModel.User.findByIdAndRemove(req.params.id);
-    payload.message = 'usuario eliminado';
-    res.send(payload);
+user.delete = async (req, res) =>{
+    try{
+        console.log(req.params.id)
+        if (hlpUser.UserExist(req.body.username) == true){
+            var usr = scModel.User.find({"username": req.body.username});
+            if (md5(req.body.password) == usr[0].username){
+                await scModel.User.findByIdAndRemove(req.params.id);
+                res.status(200);
+                res.send({message: 'usuario eliminado'})
+            }else{
+                res.status(403);
+                res.send({message: 'contraseña incorecta'})
+            }
+        }else{
+            console.log("Usuario no existente")
+            res.status(404);
+            res.send({message: 'Usisario no encontrado'})
+        }
+    }catch(err){
+        console.log(err)
+    }
 };
 
 module.exports = user;
